[IMP] Rename cost_price to 'Replenishment cost'
Fixes #6
This change is meant as a short-term solution, to avoid confusion between standard_price and cost_price since both are named 'Cost Price' in the GUI.
The accounting best practise seems to be that if a Cost price is known, it MUST be used for the valuation of the stock. So the right 'Cost price' field is standard_price and we have to name cost_price something else.
Translation into French is provided.
The description of the fields in the modules which override it is changed too, because for some reason the server seems unable to export the new names in the .po files.