[FIX] sale_margin_security: onchange product
In sale_margin, when the line already exists, the purchase_price
is computed with the product onchange method, so if the user doesn't
have the field available in the view it's not going to get the right
price. This should be fixed in v13 with the new compute fields
possibilities.